refactor: strategy 패키지 생성
